+++ 30000day/5.Other/ThirdPart/PhotoBrowser/ZLThumbnailViewController.m
@@ -118,7 +118,7 @@
 {
     if (_arraySelectPhotos.count >= self.maxSelectCount
         && btn.selected == NO) {
-        ShowToastLong(@"最多只能选择%ld张图片", self.maxSelectCount);
+        ShowToastLong(@"最多只能选择%ld张图片", (long)self.maxSelectCount);
         return;
     }
